Add 'connect' to render data from the redux store
Add connect to the App component to declare it as an entry point for
data from the Redux store.
State refers to the data in the store, and props are what get passed to
the connected component.
Whatever you return from mapStateToProps will be passed to your
component as props.

Listing 2.11. src/App.js: connecting components